chore(scan): remove unused dependencies (#193)
Removes unused dependencies from the `scan` package:

- `@babel/generator` - not used
- `@clack/core` - not used directly, already pulled in via
  `@clack/prompts`
- `estree-walker` - not used
- `tsx` - not used

At the root of the monorepo:

- `@vercel/style-guide` - doesn't seem to be referenced anywhere

Also changes a couple of FS paths to use `__dirname` in case `tsup` is
ever run from another `cwd`.

This should drop a big chunk of the install size, re #178. Specifically
since we drop `tsx` which is ~400KB.

Co-authored-by: RobPruzan <97781863+RobPruzan@users.noreply.github.com>